consolidate more generic utils
