--- src/utils/character/modal/elementColors.ts
@@ -1,11 +1,0 @@
-const elementColors = {
-    pyro: '#bf612d',
-    geo: '#c8922b',
-    dendro: '#84a02f',
-    anemo: '#5d9b86',
-    hydro: '#3d9bc1',
-    cryo: '#7fabb6',
-    electro: '#8c729a',
-}
-
-export default elementColors
