+++ src/utils/index.ts
@@ -11,7 +11,6 @@ export {
 } from './character/characterOptions'
 export { default as handleBonusToggle } from './character/handleBonusToggle'
 export { default as compareElement } from './character/modal/compareElement'
-export { default as elementColors } from './character/modal/elementColors'
 export { attributeKeyMap, fullBaseStats } from './damage/attributeKeyMapping'
 export { default as calculateDamage } from './damage/calculateDamage'
 export { default as convertBaseStats } from './damage/convertBaseStats'
@@ -20,5 +19,5 @@ export { default as genericFormulaWithScaling } from './damage/formulas/genericF
 export { default as genericFormulaWithoutScaling } from './damage/formulas/genericFormulaWithoutScaling'
 export { default as useActiveConstellations } from './hooks/useActiveConstellations'
 export { default as useBaseStats } from './hooks/useBaseStats'
-export { compareObjects, parseScalingValue } from './utils'
+export { compareObjects, parseScalingValue, elementColors } from './utils'
