+++ src/utils/utils.ts
@@ -14,4 +14,14 @@ const parseScalingValue = (scalingValue: string) => {
     return parsedValues
 }
 
-export { compareObjects, parseScalingValue }
+const elementColors = {
+    pyro: '#bf612d',
+    geo: '#c8922b',
+    dendro: '#84a02f',
+    anemo: '#5d9b86',
+    hydro: '#3d9bc1',
+    cryo: '#7fabb6',
+    electro: '#8c729a',
+}
+
+export { compareObjects, parseScalingValue, elementColors }
